optimized register controller
