--- backend/src/routes/1estudiante.routes.js
@@ -1,1 +1,0 @@
-/// RUTAS ROL ESTUDIANTE
